Quick Jump Limiting Fix
Just a tweak from a game design perspective. I had it set so that the player could eject (jump) off of parts they were detaching from, but they could do this without any jumps left. This could be a problem as it gives the player triple jump, so I have removed it for now.