Fixed box pickup so it no longer goes through walls
Previously, the box disabled its collider when the player picked it up. If a player walked it into a wall and dropped it, it would then fall out of the map. Now the player character has a secondary circlecollider2D on it, that enables itself in the boxes snap position when the box is picked up. It has a slippery material and is using a circle collider because I found this is easier to jump past ledges with than a square collider (or higher friction, which sticks to walls).