Box Assignment! :)
Boxes now use a while loop to ensure that the box closest to the holding location is always picked up when the player tries to pick up a box.
I have also changed a couple of the box prefabs so that they feel better stacking. This may need further adjustment in the future to ensure that  they work with our other systems