Lifter Arms Added, Hookshot In progress
Lifter arms are the only ones capable of carrying boxes on the HeavyLiftable layer. If you only have basic arms, and the closest box is HeavyLiftable, you will not lift any boxes.

In progress to completing the hookshot. This can be used while you are a head, and it is controlled using the right mouse button.