Time Slow
I've experienced a glitch where using floats with changing values while a time slow is applied is causing them to return two values at the same time.
I've had to use a simple slow down effect, but I will look into this with Brad when I get the chance, so that I can get the smoother transition to work consistently.