Got the lasers working
The banshee now accurately tracks the player, and fires a preliminary laser at them.
During this time, if the player has the shield, they fire a preliminary laser as well.
After 1.5 seconds, the Banshee screams, sending out death laser for 3 seconds. This kills the player if they are unshielded or in the way, and alters the player's laser if they have the shield out.
Getting out of sight will cause the laser time to wind down. The banshee should also stop rotating while firing the death laser.