Conveyor belt, attempt at elevator and fan
Conveyor belt is working perfectly. Input a force value (negative is left, positive is right) to create a zone of ground that pushes things in a certain area.

The elevator was having wierd scaling issue when the player detached on it. This is now fixed, but there are also collider problems I'm still working on.

The fan will need a lot of fine tuning, and I am still working on this.