Laser Router, Power Cells and Power Cell Stations
We now have the framework from which to build laser puzzles using routers, carriable cells, and stations, which can charge cells or act as buttons when a charged cell is placed inside.
Currently the station is not fully configured, but it can now easily be applied in the same way as a button.