Stopped the level from being affected by the elevator
Sometimes the elevator would become the parent of the fall through platforms (due to a physics collision with the Non Player Layer they use as their secondary collider, specifically if the player falls through one while an elevator also overlaps it). This would result in all of the platforms moving with the elevator. No longer a problem, the layers are now no longer colliding.