Line Animation for Banshee
Line animation is now in. I had to change my code based on a few unity answers regarding how to get the scale and offset to work together. It works very nicely now.
The next stem will be to add an entrance and exit animation, and potentially particles flying off the side of it.