Scaler Fix
Previously you could sometimes roll up walls without actually rotating, this is now fixed.
Whenever you are moving up a wall, the augment will detect if there is not horizontal input, if not, it applies a strong horizontal input towards the wall, generating friction for the angular drag to rotate the character.