Laser refinery
Laser routers are now fully implemented as both activators and activated dependent objects.
They also can now be moved around in the scene.
I have changed some parameters regarding the firing of lasers, adding a burst effect from the point of emission, and offsetting the beginning of the line renderer to avoid it from stopping inside of the laser emitter.