Camera Shifter Triggers cleanup
You can now see the camera size at each location.
Gizmos only appear when it is clicked on in the editor scene.
If you lock an axis, the cross in the middle of the camera zone will show a green line for the unlocked axis, and a red one for the locked axis).